+++ DB/DBC/standard/business/src/test/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShokanJutakuKaishuBuilderTest.java
@@ -35,7 +35,7 @@ public class ShokanJutakuKaishuBuilderTest extends DbcTestBase {
     private static JigyoshaNo 主キー名4;
     private static RString 主キー名5;
     private static RString 主キー名6;
-    private static Decimal 主キー名7;
+    private static int 主キー名7;
 
     @BeforeClass
     public static void setUpClass() {
